requirements.txt
# ===========================================
# FACE CLUSTERING PROJECT - REQUIREMENTS
# ===========================================
# Install: pip install -r requirements.txt
# ===========================================

# Core AI & Math
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Face Recognition (InsightFace)
insightface>=0.7.3
onnxruntime>=1.16.0

# Computer Vision
opencv-python-headless>=4.8.0
pillow>=10.0.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Workflow Orchestration
prefect>=2.14.0

# Data Storage
pyarrow>=14.0.0

# Vector Search
faiss-cpu>=1.7.4

# Utils
tqdm>=4.65.0
pyyaml>=6.0.0
requests>=2.31.0

# Visualization (optional - for evaluation)
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.3

# Job Queue (for background processing)
celery>=5.3.0
redis>=5.0.0

# YouTube Download
yt-dlp>=2024.1.0